+++ elevenlabs_rs/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `name` and `access_level` fields to `GetKnowledgeBaseResponse` struct
 - `ListKnowledgeBasesDocs` endpoint
 - `ListDependentAgents` endpoint
+- `DeleteKnowledgeBaseDoc` endpoint
 ### Changed
 - **Breaking**: The `ResponseBody` of `DeleteAgent` now returns a `()` instead of a `StatusResponseBody`
 - **Breaking**: The `used_tools` field name to `tool_ids` of the `PromptConfig` and its type `Option<Vec<UsedTool>>`